A proxy aware dialer introduced.
By default, if no proxy resolver function is set when creating the consumer, the 'http.ProxyFromEnvironment' proxy URL function will by used. The proxy URL function is resposible to provide the URL of the proxy server.

The URL will be used to establish a TCP connection to the proxy server following with a HTTP CONNECT request in order to establish a proxied connection to the loggregator server.

If no proxy URL is resolved or the target URL should not be proxied, dialer will fall back to 'net.Dial'.