Read entire body data for HTTP 401 responses
Noticed that a more thoughtful developer, namely @jmtuley and @ajackson,
wrote a very similar line of code in the case of HTTP fallback that doesn't
include a hardcoded buffer size. This seems like a clear win, but I was
unsure if a test for this behavior would be reasonable or just bloat.

Totally willing to write a test for this behavior.